refactor(push): extract chat notification audience
